chore(tests): optimize naming of tilde tests
